Added the code to handle much larger numbers

+++ src/com/assignment/Main.java
@@ -1,11 +1,11 @@
 package com.assignment;
 
 public class Main {
-	public int Add(String numbers) {
+	public long Add(String numbers) {
 		if(numbers.length() > 0) {
 			String[] numberArr = numbers.split(",");
-			int sum = 0;
-			for(int i = 0; i < numberArr.length; i++) sum += Integer.parseInt(numberArr[i].trim()); 
+			long sum = 0;
+			for(int i = 0; i < numberArr.length; i++) sum += Long.parseLong(numberArr[i].trim()); 
 			return sum;
 		}
 		return 0;
